Dataset: covercrops

Consider a 3 ×3 factorial experiment in randomized blocks, with 4 replications, on the influence of three new soybean cultivars (A1, A2 and A3) and the use of three types of green manure (B1, B2 and B3) on yield in 100 m2 plots.

data(covercrops)

Format

data.frame containing data set

A

Categorical vector with cultivars

B

Categorical vector with green manure

Bloco

Categorical vector with block

Resp

Numeric vector with yield

See also

Examples

data(covercrops)