Dataset: Tomato data

An experiment conducted in a randomized block design in a split plot scheme was developed in order to evaluate the efficiency of bacterial isolates in the development of tomato cultivars. The experiment counted a total of 24 trays; each block (in a total of four blocks), composed of 6 trays, in which each tray contained a treatment (6 isolates). Each tray was seeded with 4 different genotypes, each genotype occupying 28 cells per tray. The trays were randomized inside each block and the genotypes were randomized inside each tray.

data(tomate)

Format

data.frame containing data set

parc

Categorical vector with plot

subp

Categorical vector with split-plot

bloco

Categorical vector with block

resp

Numeric vector

See also

Examples

data(tomate)
attach(tomate)

#> The following object is masked from simulate1 (pos = 3):
#> 
#>     resp

#> The following object is masked from sensorial (pos = 7):
#> 
#>     resp

#> The following objects are masked from laranja (pos = 8):
#> 
#>     bloco, resp

#> The following object is masked from simulate3 (pos = 9):
#> 
#>     resp

#> The following objects are masked from simulate2 (pos = 10):
#> 
#>     bloco, resp

#> The following object is masked from simulate1 (pos = 11):
#> 
#>     resp

#> The following object is masked from sensorial (pos = 12):
#> 
#>     resp

#> The following objects are masked from laranja (pos = 13):
#> 
#>     bloco, resp

#> The following objects are masked from laranja (pos = 14):
#> 
#>     bloco, resp

#> The following objects are masked from enxofre (pos = 15):
#> 
#>     bloco, resp

#> The following objects are masked from tomate (pos = 16):
#> 
#>     bloco, parc, resp, subp

#> The following objects are masked from tomate (pos = 18):
#> 
#>     bloco, parc, resp, subp

#> The following objects are masked from cloro (pos = 22):
#> 
#>     bloco, resp

#> The following objects are masked from mirtilo (pos = 24):
#> 
#>     bloco, resp

#> The following object is masked from emerg:
#> 
#>     resp

#> The following objects are masked from enxofre (pos = 26):
#> 
#>     bloco, resp

#> The following objects are masked from enxofre (pos = 27):
#> 
#>     bloco, resp

#> The following objects are masked from cloro (pos = 29):
#> 
#>     bloco, resp

#> The following objects are masked from cloro (pos = 30):
#> 
#>     bloco, resp

#> The following objects are masked from cloro (pos = 31):
#> 
#>     bloco, resp

#> The following objects are masked from cloro (pos = 33):
#> 
#>     bloco, resp

#> The following objects are masked from cloro (pos = 34):
#> 
#>     bloco, resp

#> The following objects are masked from cloro (pos = 35):
#> 
#>     bloco, resp

#> The following objects are masked from cloro (pos = 36):
#> 
#>     bloco, resp

#> The following object is masked from simulate3 (pos = 37):
#> 
#>     resp

#> The following object is masked from porco:
#> 
#>     resp

#> The following object is masked from simulate1 (pos = 39):
#> 
#>     resp

#> The following object is masked from aristolochia (pos = 43):
#> 
#>     resp

#> The following objects are masked from enxofre (pos = 44):
#> 
#>     bloco, resp

#> The following objects are masked from cloro (pos = 45):
#> 
#>     bloco, resp

#> The following objects are masked from simulate2 (pos = 47):
#> 
#>     bloco, resp

#> The following objects are masked from laranja (pos = 49):
#> 
#>     bloco, resp

#> The following object is masked from aristolochia (pos = 50):
#> 
#>     resp

#> The following objects are masked from mirtilo (pos = 51):
#> 
#>     bloco, resp

#> The following objects are masked from mirtilo (pos = 52):
#> 
#>     bloco, resp

#> The following objects are masked from laranja (pos = 53):
#> 
#>     bloco, resp

#> The following object is masked from passiflora:
#> 
#>     bloco