Analysis: DBC experiments in split-split-plot

Analysis of an experiment conducted in a randomized block design in a split-split-plot scheme using analysis of variance of fixed effects.

PSUBSUBDBC(
  f1,
  f2,
  f3,
  block,
  response,
  alpha.f = 0.05,
  alpha.t = 0.05,
  dec = 3,
  mcomp = "tukey"
)

Arguments

f1
Numeric or complex vector with plot levels

f2
Numeric or complex vector with splitplot levels

f3
Numeric or complex vector with splitsplitplot levels

block
Numeric or complex vector with blocks

response
Numeric vector with responses

alpha.f
Level of significance of the F test (default is 0.05)

alpha.t
Significance level of the multiple comparison test (default is 0.05)

dec
Number of cells

mcomp
Multiple comparison test (Tukey (default), LSD and Duncan)

Value

Analysis of variance of fixed effects and multiple comparison test of Tukey, LSD or Duncan.

Note

The PSUBSUBDBC function does not present residual analysis, interaction breakdown, graphs and implementations of various multiple comparison or regression tests. The function only returns the analysis of variance and multiple comparison test of Tukey, LSD or Duncan.

Author

Gabriel Danilo Shimizu, shimizu@uel.br

Leandro Simoes Azeredo Goncalves

Rodrigo Yudi Palhaci Marubayashi

Examples

library(AgroR)
data(enxofre)
with(enxofre, PSUBSUBDBC(f1, f2, f3, bloco, resp))

#> 
#> -----------------------------------------------------------------
#> Normality of errors
#> -----------------------------------------------------------------
#>                          Method Statistic   p.value
#>  Shapiro-Wilk normality test(W) 0.9864607 0.3476439
#> 

#> As the calculated p-value is greater than the 5% significance level, hypothesis H0 is not rejected. Therefore, errors can be considered normal

#> 
#> 
#> -----------------------------------------------------------------
#> Homogeneity of Variances
#> -----------------------------------------------------------------
#>                               Method Statistic   p.value
#>  Bartlett test(Bartlett's K-squared)  20.01974 0.7906199
#> 

#> As the calculated p-value is greater than the 5% significance level, hypothesis H0 is not rejected. Therefore, the variances can be considered homogeneous

#> Warning: Error() model is singular

#> 
#> -----------------------------------------------------------------
#> Additional Information
#> -----------------------------------------------------------------
#> 
#> CV plot (%) =  5.46
#> CV split plot (%) =  7.42
#> CV split split plot (%) =  6.53
#> Mean =  4985.6044
#> Median =  4975.775
#> 
#> -----------------------------------------------------------------
#> Analysis of Variance
#> -----------------------------------------------------------------
#>                      Df    Sum Sq    Mean Sq F value  Pr(>F)
#> Fator1                2  532880.7  266440.37 3.59743 0.09402
#> Residuals             6  444385.1   74064.18                
#> Fator2                2 2593572.1 1296786.07 9.48772 0.00154
#> Fator1:Fator2         4  568195.7  142048.92 1.03928 0.41434
#> Residuals            18 2460248.8  136680.49                
#> Fator3                2 1012836.0  506417.98  4.7814 0.01225
#> Fator1:Fator3         4 2177620.9  544405.23 5.14006 0.00139
#> Fator2:Fator3         4  548172.3  137043.08 1.29391 0.28399
#> Fator1:Fator2:Fator3  8 2255321.1  281915.14 2.66173 0.01537
#> Residuals            54 5719361.9  105914.11                
#> 
#> ------------------------------------------
#> Interaction F1*F2*F3  significant: unfolding the interaction
#> ------------------------------------------
#> 
#> ------------------------------------------
#> Analyzing  F1  inside of each level of  F2 and F3
#> ------------------------------------------
#>                                            Df      Sum sq   Mean Sq   F value
#> Fator3:Fator1:Fator2                       18 5965261.253 331403.40 3.0535813
#>   Fator3:Fator1:Fator2: 75 Vegetativo       2  236015.402 118007.70 1.0158228
#>   Fator3:Fator1:Fator2: 225 Vegetativo      2  835403.885 417701.94 3.5956227
#>   Fator3:Fator1:Fator2: 675 Vegetativo      2  935907.398 467953.70 4.0281952
#>   Fator3:Fator1:Fator2: 75 Ciclo Completo   2 1320014.218 660007.11 5.6814114
#>   Fator3:Fator1:Fator2: 225 Ciclo Completo  2  136069.198  68034.60 0.5856491
#>   Fator3:Fator1:Fator2: 675 Ciclo Completo  2  486225.501 243112.75 2.0927404
#>   Fator3:Fator1:Fator2: 75 Reprodutivo      2 1859098.181 929549.09 8.0016574
#>   Fator3:Fator1:Fator2: 225 Reprodutivo     2   53620.778  26810.39 0.2307867
#>   Fator3:Fator1:Fator2: 675 Reprodutivo     2  102906.693  51453.35 0.4429159
#> Residuals combined                         65    1787.224 116169.57          
#>                                                  Pr(>F)
#> Fator3:Fator1:Fator2                       0.0003129384
#>   Fator3:Fator1:Fator2: 75 Vegetativo      0.3677798511
#>   Fator3:Fator1:Fator2: 225 Vegetativo     0.0330321405
#>   Fator3:Fator1:Fator2: 675 Vegetativo     0.0224280876
#>   Fator3:Fator1:Fator2: 75 Ciclo Completo  0.0053213269
#>   Fator3:Fator1:Fator2: 225 Ciclo Completo 0.5596549076
#>   Fator3:Fator1:Fator2: 675 Ciclo Completo 0.1315823918
#>   Fator3:Fator1:Fator2: 75 Reprodutivo     0.0007822460
#>   Fator3:Fator1:Fator2: 225 Reprodutivo    0.7945565540
#>   Fator3:Fator1:Fator2: 675 Reprodutivo    0.6440847567
#> Residuals combined                                     
#> 
#> 
#> ------------------------------------------
#>  F1  within the combination of levels  150  of   F2  and  Vegetativo  of   F3 
#> ------------------------------------------
#>         Mean letters
#> 75  5246.363       a
#> 675 4578.007       b
#> 225 4429.288       b
#> 
#> 
#> ------------------------------------------
#>  F1  within the combination of levels  150  of   F2  and  Ciclo Completo  of   F3 
#> ------------------------------------------
#>         Mean letters
#> 675 5027.233       a
#> 225 4927.000       a
#> 75  4792.932       a
#> 
#> 
#> ------------------------------------------
#>  F1  within the combination of levels  150  of   F2  and  Reprodutivo  of   F3 
#> ------------------------------------------
#>         Mean letters
#> 675 5056.610       a
#> 225 4847.748      ab
#> 75  4335.730       b
#> 
#> 
#> ------------------------------------------
#>  F1  within the combination of levels  450  of   F2  and  Vegetativo  of   F3 
#> ------------------------------------------
#>         Mean letters
#> 75  5286.368       a
#> 225 5125.075       a
#> 675 4472.670       b
#> 
#> 
#> ------------------------------------------
#>  F1  within the combination of levels  450  of   F2  and  Ciclo Completo  of   F3 
#> ------------------------------------------
#>         Mean letters
#> 75  5389.980       a
#> 675 5085.130       a
#> 225 4889.233       a
#> 
#> 
#> ------------------------------------------
#>  F1  within the combination of levels  450  of   F2  and  Reprodutivo  of   F3 
#> ------------------------------------------
#>         Mean letters
#> 675 4968.540       a
#> 75  4847.222       a
#> 225 4658.573       a
#> 
#> 
#> ------------------------------------------
#>  F1  within the combination of levels  1350  of   F2  and  Vegetativo  of   F3 
#> ------------------------------------------
#>         Mean letters
#> 75  5561.840       a
#> 675 5355.302       a
#> 225 5140.382       a
#> 
#> 
#> ------------------------------------------
#>  F1  within the combination of levels  1350  of   F2  and  Ciclo Completo  of   F3 
#> ------------------------------------------
#>         Mean letters
#> 75  5305.762       a
#> 675 5188.823       a
#> 225 5131.625       a
#> 
#> 
#> ------------------------------------------
#>  F1  within the combination of levels  1350  of   F2  and  Reprodutivo  of   F3 
#> ------------------------------------------
#>         Mean letters
#> 225 5147.488       a
#> 75  4984.855       a
#> 675 4831.542       a
#> 
#> 
#> 
#> ------------------------------------------
#> Analyzing  F2  inside of each level of  F1 and F3
#> ------------------------------------------
#>                                             Df      Sum sq   Mean Sq   F value
#> Fator3:Fator2:Fator1                        18 5534018.435 307445.47 2.8328307
#>   Fator3:Fator2:Fator1: 150 Vegetativo       2 1515236.800 757618.40 7.9500370
#>   Fator3:Fator2:Fator1: 450 Vegetativo       2  110556.185  55278.09 0.5800584
#>   Fator3:Fator2:Fator1: 1350 Vegetativo      2 1100604.585 550302.29 5.7745741
#>   Fator3:Fator2:Fator1: 150 Ciclo Completo   2 1485001.568 742500.78 7.7914010
#>   Fator3:Fator2:Fator1: 450 Ciclo Completo   2  509409.882 254704.94 2.6727357
#>   Fator3:Fator2:Fator1: 1350 Ciclo Completo  2  195182.146  97591.07 1.0240678
#>   Fator3:Fator2:Fator1: 150 Reprodutivo      2  355299.693 177649.85 1.8641612
#>   Fator3:Fator2:Fator1: 450 Reprodutivo      2   63027.182  31513.59 0.3306866
#>   Fator3:Fator2:Fator1: 1350 Reprodutivo     2  199700.394  99850.20 1.0477739
#> Residuals combined                          65    2027.606  95297.47          
#>                                                   Pr(>F)
#> Fator3:Fator2:Fator1                        0.0007435565
#>   Fator3:Fator2:Fator1: 150 Vegetativo      0.0010616894
#>   Fator3:Fator2:Fator1: 450 Vegetativo      0.5638228728
#>   Fator3:Fator2:Fator1: 1350 Vegetativo     0.0057222390
#>   Fator3:Fator2:Fator1: 150 Ciclo Completo  0.0011956583
#>   Fator3:Fator2:Fator1: 450 Ciclo Completo  0.0795500520
#>   Fator3:Fator2:Fator1: 1350 Ciclo Completo 0.3670040040
#>   Fator3:Fator2:Fator1: 150 Reprodutivo     0.1663079024
#>   Fator3:Fator2:Fator1: 450 Reprodutivo     0.7200882693
#>   Fator3:Fator2:Fator1: 1350 Reprodutivo    0.3587650124
#> Residuals combined                                      
#> 
#> 
#> ------------------------------------------
#>  F2  within the combination of levels  75  of   F1  and  Vegetativo  of   F3 
#> ------------------------------------------
#>          Mean letters
#> 1350 5561.840       a
#> 450  5286.368       a
#> 150  5246.363       a
#> 
#> 
#> ------------------------------------------
#>  F2  within the combination of levels  75  of   F1  and  Ciclo Completo  of   F3 
#> ------------------------------------------
#>          Mean letters
#> 450  5389.980       a
#> 1350 5305.762      ab
#> 150  4792.932       b
#> 
#> 
#> ------------------------------------------
#>  F2  within the combination of levels  75  of   F1  and  Reprodutivo  of   F3 
#> ------------------------------------------
#>          Mean letters
#> 1350 4984.855       a
#> 450  4847.222      ab
#> 150  4335.730       b
#> 
#> 
#> ------------------------------------------
#>  F2  within the combination of levels  225  of   F1  and  Vegetativo  of   F3 
#> ------------------------------------------
#>          Mean letters
#> 1350 5140.382       a
#> 450  5125.075       a
#> 150  4429.288       b
#> 
#> 
#> ------------------------------------------
#>  F2  within the combination of levels  225  of   F1  and  Ciclo Completo  of   F3 
#> ------------------------------------------
#>          Mean letters
#> 1350 5131.625       a
#> 150  4927.000       a
#> 450  4889.233       a
#> 
#> 
#> ------------------------------------------
#>  F2  within the combination of levels  225  of   F1  and  Reprodutivo  of   F3 
#> ------------------------------------------
#>          Mean letters
#> 1350 5147.488       a
#> 150  4847.748       a
#> 450  4658.573       a
#> 
#> 
#> ------------------------------------------
#>  F2  within the combination of levels  675  of   F1  and  Vegetativo  of   F3 
#> ------------------------------------------
#>          Mean letters
#> 1350 5355.302       a
#> 150  4578.007       b
#> 450  4472.670       b
#> 
#> 
#> ------------------------------------------
#>  F2  within the combination of levels  675  of   F1  and  Ciclo Completo  of   F3 
#> ------------------------------------------
#>          Mean letters
#> 1350 5188.823       a
#> 450  5085.130       a
#> 150  5027.233       a
#> 
#> 
#> ------------------------------------------
#>  F2  within the combination of levels  675  of   F1  and  Reprodutivo  of   F3 
#> ------------------------------------------
#>          Mean letters
#> 150  5056.610       a
#> 450  4968.540       a
#> 1350 4831.542       a
#> 
#> ------------------------------------------
#> Analyzing  F3  inside of each level of  F1 and F2
#> ------------------------------------------
#>                                  Df       Sum sq     Mean Sq     F value
#> Fator1:Fator2:Fator3             18 5993950.3194 332997.2400 3.144030966
#>   Fator1:Fator2:Fator3: 75 150    2 1658512.5879 829256.2940 7.829516747
#>   Fator1:Fator2:Fator3: 225 150   2  572143.2840 286071.6420 2.700977645
#>   Fator1:Fator2:Fator3: 675 150   2  575635.3215 287817.6608 2.717462878
#>   Fator1:Fator2:Fator3: 75 450    2  664226.1133 332113.0567 3.135682850
#>   Fator1:Fator2:Fator3: 225 450   2  435267.0706 217633.5353 2.054811548
#>   Fator1:Fator2:Fator3: 675 450   2  846116.7155 423058.3577 3.994353158
#>   Fator1:Fator2:Fator3: 75 1350   2  668625.3330 334312.6665 3.156450714
#>   Fator1:Fator2:Fator3: 225 1350  2     505.0583    252.5292 0.002384283
#>   Fator1:Fator2:Fator3: 675 1350  2  572918.8352 286459.4176 2.704638872
#> Residuals                        54 5719361.9122 105914.1095            
#>                                        Pr(>F)
#> Fator1:Fator2:Fator3             0.0005864037
#>   Fator1:Fator2:Fator3: 75 150   0.0010333335
#>   Fator1:Fator2:Fator3: 225 150  0.0762099225
#>   Fator1:Fator2:Fator3: 675 150  0.0750766616
#>   Fator1:Fator2:Fator3: 75 450   0.0514794196
#>   Fator1:Fator2:Fator3: 225 450  0.1380174292
#>   Fator1:Fator2:Fator3: 675 450  0.0241097134
#>   Fator1:Fator2:Fator3: 75 1350  0.0505307264
#>   Fator1:Fator2:Fator3: 225 1350 0.9976186627
#>   Fator1:Fator2:Fator3: 675 1350 0.0759567119
#> Residuals                                    
#> 
#> 
#> ------------------------------------------
#>  F3  within the combination of levels  75  of  F1  and  150  of   F2 
#> ------------------------------------------
#>                    Mean letters
#> Vegetativo     5246.363       a
#> Ciclo Completo 4792.932      ab
#> Reprodutivo    4335.730       b
#> 
#> 
#> ------------------------------------------
#>  F3  within the combination of levels  75  of  F1  and  450  of   F2 
#> ------------------------------------------
#>                    Mean letters
#> Ciclo Completo 5389.980       a
#> Vegetativo     5286.368       a
#> Reprodutivo    4847.222       a
#> 
#> 
#> ------------------------------------------
#>  F3  within the combination of levels  75  of  F1  and  1350  of   F2 
#> ------------------------------------------
#>                    Mean letters
#> Vegetativo     5561.840       a
#> Ciclo Completo 5305.762      ab
#> Reprodutivo    4984.855       b
#> 
#> 
#> ------------------------------------------
#>  F3  within the combination of levels  225  of  F1  and  150  of   F2 
#> ------------------------------------------
#>                    Mean letters
#> Ciclo Completo 4927.000       a
#> Reprodutivo    4847.748       a
#> Vegetativo     4429.288       a
#> 
#> 
#> ------------------------------------------
#>  F3  within the combination of levels  225  of  F1  and  450  of   F2 
#> ------------------------------------------
#>                    Mean letters
#> Vegetativo     5125.075       a
#> Ciclo Completo 4889.233       a
#> Reprodutivo    4658.573       a
#> 
#> 
#> ------------------------------------------
#>  F3  within the combination of levels  225  of  F1  and  1350  of   F2 
#> ------------------------------------------
#>                    Mean letters
#> Reprodutivo    5147.488       a
#> Vegetativo     5140.382       a
#> Ciclo Completo 5131.625       a
#> 
#> 
#> ------------------------------------------
#>  F3  within the combination of levels  675  of  F1  and  150  of   F2 
#> ------------------------------------------
#>                    Mean letters
#> Reprodutivo    5056.610       a
#> Ciclo Completo 5027.233       a
#> Vegetativo     4578.007       a
#> 
#> 
#> ------------------------------------------
#>  F3  within the combination of levels  675  of  F1  and  450  of   F2 
#> ------------------------------------------
#>                   Mean letters
#> Ciclo Completo 5085.13       a
#> Reprodutivo    4968.54      ab
#> Vegetativo     4472.67       b
#> 
#> 
#> ------------------------------------------
#>  F3  within the combination of levels  675  of  F1  and  1350  of   F2 
#> ------------------------------------------
#>                    Mean letters
#> Vegetativo     5355.302       a
#> Ciclo Completo 5188.823       a
#> Reprodutivo    4831.542       a